Fixed MessageDetailsActivityTest. Refactored code.| #956
