+++ FlowCrypt/src/androidTest/assets/messages/mime/8bit-utf8.txt
@@ -1,67 +1,20 @@
-Return-Path: <denbond7@denbond7.com>
-Delivered-To: default@denbond7.com
-Received: from mail.denbond7.com
-	by mail.denbond7.com with LMTP
-	id 3yVQLfBPe1+6QAAAZlazJA
-	(envelope-from <denbond7@denbond7.com>)
-	for <default@denbond7.com>; Mon, 05 Oct 2020 16:55:12 +0000
-Received: from localhost (localhost [127.0.0.1])
-	by mail.denbond7.com (Postfix) with ESMTP id B623BB20B92
-	for <default@denbond7.com>; Mon,  5 Oct 2020 16:55:12 +0000 (UTC)
-To: default@denbond7.com
-From: Denis Bondarenko <denbond7@denbond7.com>
+Date: Mon, 29 Jul 2019 07:00:45 +0000
+To: 7bitencoding@domain.com
+From: "OLX.ua" <noreply@olx.ua>
 Subject: 8bit encoded utf broken
-Autocrypt: addr=denbond7@denbond7.com; keydata=
- mQENBFyM4vkBCACr8FvuG4MAZsz90hrZ2B+oEznArD46QUbAGrPT8JU6zz27SfLmTTbTA86q
- +DVAZC94CZ2lqJaG7KY7eqHFqJObetA8mcUk0Oo3EshphpK0MnNv+bHFpKf/srJgKSDmpCF+
- JSHeZs2WAPtrdfoTO1KPy5IZexMWdnhYZy8d0xUr4rPZ7i2UyrkdV3k2O9x4aG5tD1Zo/Wkm
- lTv8z2BtAwAziqUQFxABQLu1fzcRkMHTMcGVXrB+73GITl03GbfeJHvFuThhofqP09o/p3Y1
- 8Zm+zVWL/T4zsA5P+Cp6kwnwUnQkL3Tc49a0qIJShlP4LbSbPBqFGWTHxceCkFBkMVoJABEB
- AAG0LWRlbmJvbmQ3QGRlbmJvbmQ3LmNvbSA8ZGVuYm9uZDdAZGVuYm9uZDcuY29tPokBPwQQ
- AQgAKQUCXIzjAQYLCQcIAwIJEGoVBlPxgAh0BBUICgIDFgIBAhkBAhsDAh4BAAoJEGoVBlPx
- gAh0L2EIAJRM22lnSHhgQ+waiCf0p739L3XKs2eDZUkxAX+4an654wnn1OjxWrG7lls2GgN1
- ziKHHdGgxU4PQ8PmQrgyc6F05vmYihRFaVFDQ97xF16+IvO4xiuYMuJk7ovsr5mrlI43CWE4
- f7GPtmJ7QYTlxeYW3s7KLqr0TKxqGG8f6qRjy/Q1ky3qnseR5Fp1GjIwijVI2rncfH0lOkNm
- BHj4aYugN60HNlUJbosb2vjJkz5dsqe+nY3XVKI6Iqq07I+Jvankn6jyvlMP8jNLxQMUsqrt
- Bki/C6ett3z+JeP2+2Nh+mp1vDW/kO7gSgRHUEPJ9w/VVIpWIOTEv+Bsz3l28h65AQ0EXIzj
- AAEIAJKbKexaP/2HqUAj/DxOebhKqw2ddK5HsYacv12pHVTEVjOnLQELyroZ1ikO21UK8Z/z
- ng9+8DlCz72SMql+/uZYpoEdgqMy9PU3zPDzGcAeOr07R56G08l+hUO/i3FMW4v2ZT6gLCIn
- AJ417wSvF4wg6dJzfrBFkj7o2lyTOwfD87CehQozyGza1cb+yS3ZPC8oc99a1jwSqFYecDI2
- 7Cgy7KWSbhSfoPcJG4KPTLDrvVn2Y+U2jxx6OiYXRai8wwlheAmIohmZNpHLaCQiTZTWRYPK
- vP0kb6aoN9O4fGhgMh3QAbc2BVwh299rToBieI/zow+I9OcrScejwQNI2VUAEQEAAYkBKQQY
- AQgAEwUCXIzjAgkQahUGU/GACHQCGwwACgkQahUGU/GACHSfcQf/euBJSXopDZxAVniN9HDQ
- Dga+ccwJYWZvyiDOaGSoSxW3QDhTOWbCH66GkZiOp9KLObHs54KzoOGaOJbQmhWGR4rwNdGS
- np8oWimyyXVEHHhir2BmT8MZrMC3WpMPmUlY2v9CMFD7hF2OfmXd1M9fuFT41aZQVDgGSdYc
- IQchOmUv3NQUiBJyOlp1bF1RT7Yf3UjAQANyENz19fYHKd2yh9Ltv3CYVjwRB4crQmuESxM8
- 7IEXDVNnYOBP0F5WaSQEesFAhjk+cVjxTUfcmM04YEFhedBStsV8kWF50ojYgF9te4yRM/zo
- NRi6SjWLSeEw/zIfvkggDvn1UfA2bGFk8Q==
-Message-ID: <6e66bc34-0e1f-bb13-ccdc-8e61bdf63f6a@denbond7.com>
-Date: Mon, 5 Oct 2020 19:55:12 +0300
-Mime-Version: 1.0
-Content-Type: multipart/alternative;
- boundary="------------FD1FA5F98C2FD885200F7353"
-Content-Language: en-US
-
-This is a multi-part message in MIME format.
---------------FD1FA5F98C2FD885200F7353
-Content-Type: text/plain; charset=utf-8
-Content-Transfer-Encoding: 8bit
-
-Text: Ð’Ð°ÑˆÐµ Ð¾Ð³Ð¾Ð»Ð¾ÑˆÐµÐ½Ð½Ñ
-
-
---------------FD1FA5F98C2FD885200F7353
+MIME-Version: 1.0
 Content-Type: text/html; charset=utf-8
 Content-Transfer-Encoding: 8bit
 
+<!DOCTYPE html>
 <html>
-  <head>
-
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-  </head>
-  <body>
-    <pre style="background-color:#ffffff;color:#000000;font-family:'DejaVu Sans Mono';font-size:12.0pt;">Text: Ð’Ð°ÑˆÐµ Ð¾Ð³Ð¾Ð»Ð¾ÑˆÐµÐ½Ð½Ñ</pre>
-  </body>
+<head>
+    <title>OLXEU</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+</head>
+<body>
+Text: Ваше оголошення
+</body>
 </html>
 
---------------FD1FA5F98C2FD885200F7353--
