+++ FlowCrypt/src/androidTest/java/com/flowcrypt/email/ui/activity/MessageDetailsActivityTest.kt
@@ -34,6 +34,7 @@ import androidx.test.espresso.web.webdriver.DriverAtoms.getText
 import androidx.test.espresso.web.webdriver.Locator
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
+import com.flowcrypt.email.DoesNotNeedMailserver
 import com.flowcrypt.email.R
 import com.flowcrypt.email.TestConstants
 import com.flowcrypt.email.api.email.EmailUtil
@@ -76,7 +77,6 @@ import java.util.concurrent.TimeUnit
  */
 @MediumTest
 @RunWith(AndroidJUnit4::class)
-//todo-denbond7 Fix storing encrypted version of message
 class MessageDetailsActivityTest : BaseTest() {
   override val useIntents: Boolean = true
   override val activeActivityRule = lazyActivityScenarioRule<MessageDetailsActivity>(launchActivity = false)
@@ -111,6 +111,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testReplyButton() {
     testStandardMsgPlaneText()
     onView(withId(R.id.layoutReplyButton))
@@ -120,11 +121,13 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testTopReplyButton() {
     testTopReplyAction(getResString(R.string.reply))
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testReplyAllButton() {
     testStandardMsgPlaneText()
     onView(withId(R.id.layoutReplyAllButton))
@@ -134,6 +137,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testFwdButton() {
     testStandardMsgPlaneText()
     onView(withId(R.id.layoutFwdButton))
@@ -143,29 +147,34 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testTopForwardButton() {
     testTopReplyAction(getResString(R.string.forward))
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testStandardMsgPlaneText() {
     baseCheck(getMsgInfo("messages/info/standard_msg_info_plain_text.json",
         "messages/mime/standard_msg_info_plain_text.txt"))
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testStandardMsgPlaneTextWithOneAttachment() {
     baseCheckWithAtt(getMsgInfo("messages/info/standard_msg_info_plain_text_with_one_att.json",
         "messages/mime/standard_msg_info_plain_text_with_one_att.txt", simpleAttInfo), simpleAttInfo)
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testEncryptedMsgPlaneText() {
     baseCheck(getMsgInfo("messages/info/encrypted_msg_info_plain_text.json",
         "messages/mime/encrypted_msg_info_plain_text.txt"))
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testEncryptedBigInlineAtt() {
     IdlingPolicies.setIdlingResourceTimeout(3, TimeUnit.MINUTES)
     baseCheck(getMsgInfo("messages/info/encrypted_msg_big_inline_att.json",
@@ -173,6 +182,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testMissingKeyErrorImportKey() {
     testMissingKey(getMsgInfo("messages/info/encrypted_msg_info_plain_text_with_missing_key.json",
         "messages/mime/encrypted_msg_info_plain_text_with_missing_key.txt"))
@@ -200,12 +210,14 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testMissingPubKey() {
     testMissingKey(getMsgInfo("messages/info/encrypted_msg_info_plain_text_error_one_pub_key.json",
         "messages/mime/encrypted_msg_info_plain_text_error_one_pub_key.txt"))
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testBadlyFormattedMsg() {
     val msgInfo = getMsgInfo("messages/info/encrypted_msg_info_plain_text_error_badly_formatted.json",
         "messages/mime/encrypted_msg_info_plain_text_error_badly_formatted.txt")
@@ -231,6 +243,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testMissingKeyErrorChooseSinglePubKey() {
     val msgInfo = getMsgInfo("messages/info/encrypted_msg_info_plain_text_with_missing_key.json",
         "messages/mime/encrypted_msg_info_plain_text_with_missing_key.txt")
@@ -250,6 +263,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testMissingKeyErrorChooseFromFewPubKeys() {
     val msgInfo = getMsgInfo("messages/info/encrypted_msg_info_plain_text_with_missing_key.json",
         "messages/mime/encrypted_msg_info_plain_text_with_missing_key.txt")
@@ -278,6 +292,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testEncryptedMsgPlaneTextWithOneAttachment() {
     val msgInfo = getMsgInfo("messages/info/encrypted_msg_info_plain_text_with_one_att.json",
         "messages/mime/encrypted_msg_info_plain_text_with_one_att.txt", encryptedAttInfo)
@@ -285,6 +300,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun testEncryptedMsgPlaneTextWithPubKey() {
     val msgInfo = getMsgInfo("messages/info/encrypted_msg_info_plain_text_with_pub_key.json",
         "messages/mime/encrypted_msg_info_plain_text_with_pub_key.txt", pubKeyAttInfo)
@@ -329,6 +345,7 @@ class MessageDetailsActivityTest : BaseTest() {
   }
 
   @Test
+  @DoesNotNeedMailserver
   fun test8bitEncodingUtf8() {
     baseCheck(getMsgInfo("messages/info/msg_info_8bit-utf8.json",
         "messages/mime/8bit-utf8.txt"))
